refactor: changed folder name docs: added .env file
