--- src/Contexts/AuthContext/AuthContext.js
@@ -1,10 +1,0 @@
-import React, { createContext, useContext, useState } from "react";
-
-const AuthContext = createContext();
-
-export const AuthContextProvider = ({ children }) => {
-  const [isLoggedIn, setIsLoggedIn] = useState(false);
-  return <AuthContext.Provider value={{ isLoggedIn, setIsLoggedIn }}>{children}</AuthContext.Provider>
-}
-
-export const AuthState = () => useContext(AuthContext);
