+++ src/Pages/ProductListing/ProductListing.js
@@ -1,6 +1,6 @@
 import React, { useEffect } from 'react'
 import { Filters } from './Filters/Filters'
-import { DataState } from '../../Contexts/DataContext/DataContext'
+import { DataState } from '../../Contexts/Data/DataContext'
 import { SingleProduct } from '../../Components/SingleProdcut/SingleProduct'
 import './ProductListing.css';
