+++ src/Reducers/AuthReducer.js
@@ -1,0 +1,12 @@
+export const authInitialState = {
+  encodedToken: "",
+  isLoggedIn: false
+}
+
+// export const authReducer = (state, action) => {
+//   switch (action.type) {
+//     case "LOGIN_AS_GUEST": {
+
+//     }
+//   }
+// }
